% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calc_GDD}
\alias{calc_GDD}
\title{Calculate growing degree days}
\usage{
calc_GDD(T_max, T_min, T_base, T_upper = NULL)
}
\arguments{
\item{T_max}{Daily Maximum Temperature}

\item{T_min}{Daily Minimum Temperature}

\item{T_base}{Base Temperature}

\item{T_upper}{If T_max is above the T_upper limit, ignore the entire day (i.e. GDD equals zero, no growth happens that day)}
}
\value{
The GDD for the given data per the formula: \code{GDD = (T_max - T_min) / 2 - T_base}
}
\description{
Calculates the daily GDD from daily maximum (T_max) and daily minimum (T_min) temperatures,
with respect to a base temperature T_base.
}
\details{
The formula for GDD is the averages of max and min temperatures minus the base temperature.
Results with negative values are set to zero (rather than having a 'negative' GDD).

The T_upper parameter allows an upper limit to be specified. If T_max is greater than T_upper, then the GDD for the day will be zero (i.e. no growth happens on that day)
}
\examples{
calc_GDD(T_max = 24.12, T_min = 15.94, T_base = 10)
}
