% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calc_GDD_weighted}
\alias{calc_GDD_weighted}
\title{Calculate Weighted Growing Degree Days}
\usage{
calc_GDD_weighted(
  T_max,
  T_min,
  T_budb = 4,
  T_leaf = 7,
  T_vera = 10,
  GDD_budb_threshold = 350,
  GDD_leaf_threshold = 1000,
  GDD_harvest_threshold = 1300,
  T_upper = 1000
)
}
\arguments{
\item{T_max}{Daily Maximum Temperature}

\item{T_min}{Daily Minimum Temperature}

\item{T_budb}{Budbreak temperature}

\item{T_leaf}{Leaf appearance temperature}

\item{T_vera}{Veraison temperature}

\item{GDD_budb_threshold}{GDD threshold}

\item{GDD_leaf_threshold}{GDD threshold}

\item{GDD_harvest_threshold}{GDD threshold}

\item{T_upper}{If T_max is above the T_upper limit, ignore the entire day (i.e. GDD equals zero, no growth happens that day)}
}
\value{
Vector of weighted GDD values
}
\description{
Calculates the cumulative weighted GDD where:
\itemize{
  \item{\code{GDD_base_budb} is used when cumulative GDD (x) meets: \code{0 < x < params[["GDD_budb_threshold"]]}}
  \item{\code{GDD_base_leaf} is used when cumulative GDD (x) meets: \code{params[["GDD_budb_threshold"]] < x < params[["GDD_leaf_threshold"]]}}
  \item{\code{GDD_base_10} is used when cumulative GDD (x) meets: \code{x > params[["GDD_leaf_threshold"]]}}
}
}
