% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calc_heatwave_EHF_severity}
\alias{calc_heatwave_EHF_severity}
\title{Calculate EHF severity}
\usage{
calc_heatwave_EHF_severity(EHF, baseline_EHF_85th, is_heatwave)
}
\arguments{
\item{EHF}{Vector of EHF values}

\item{baseline_EHF_85th}{Baseline EHF value (85th percentile)}

\item{is_heatwave}{Vector of logical values defining if corresponding EHF value is a heatwave}
}
\description{
Calculate EHF severity
}
